Profiling the naive gaussian kernel with a filter size of 6

+++ sofia/cuda/histogram.c
@@ -9,7 +9,7 @@ void histogram(double* ary, double* bins, int* histo, int ary_len, int bin_len, 
     double bin_size = bins[1] - bins[0];
 
     for (int i = 0; i < ary_len; i++) {
-        int bin = (ary[i] - min) / bin_size;
+        int bin = ((double)ary[i] - min) / bin_size;
         if (bin >= 0 && bin < histo_len) {
             histo[bin]++;
         }
